Pipe `EncodedProfile` object directly to `HttpTransport`
This change is a bit noisy since the `pprof_data` was referred in
a bunch of places, but it's mostly about switching one argument for
the other.

I explicitly decided to rename the argument to a) reflect it's no
longer pprof data that's being shipped, but our new object and
b) easily catch any leftover code/assumptions about the old argument.

This step finishes the refactoring to get the `EncodedProfile` to
`HttpTransport`. While with libdatadog 16 we had more flexibility
in how we did this, with 17 we actually need to hand libdatadog
the `ddog_prof_EncodedProfile` instead of the profile bytes.