Remove extra indirections during reporting of profiles
All this indirection was somewhat of a vestige of my first forays into
native extensions, and they created a lot of boilerplate whenever we
want to change something in the flush object, because there's now
many places to change, rather than just adding it to the flush, and
retrieving it from the flush.

The timespecs are on purpose not touched as they will be refactored
once we move to libdatadog 17, so I decided to not spend any time
on them for now.