[PROF-11524] Refactor: Avoid reading profile start_time
**What does this PR do?**

This PR changes the `StackRecorder` so that instead of relying on
libdatadog to track a profile's `start_time` (by passing `NULL` --
aka "use current timestamp") it now explicitly gets the current
timestamp and passes it along to libdatadog.

This then means that we no longer need to read the `start_time`
from libdatadog, because we're tracking it on our side.

**Motivation:**

Starting with libdatadog 17, we'll no longer be able to read
the `start_time` from libdatadog, and this PR leaves us ready for that
change.

**Additional Notes:**

N/A

**How to test the change?**

We had some existing test coverage, and I've added a bit more;
I've also injected some bugs locally on purpose to validate that the
test coverage is good. You can try doing the same ;)