[NO-TICKET] Fix profiling flaky tests due to time conversions, second try
**What does this PR do?**

This PR fixes flaky tests such as
[this one](https://github.com/DataDog/dd-trace-rb/actions/runs/14359324310/job/40256546138?pr=4570):

```
Failures:

  1) Datadog::Profiling::Collectors::ThreadContext#sample timeline support when timeline is enabled when thread starts Waiting for GVL records a second sample to represent the time spent Waiting for GVL
     Failure/Error:
       expect(second_sample.labels).to include(
         state: "waiting for gvl",
         end_timestamp_ns: be_between(time_before_sample, time_after_sample),
       )

       expected {:end_timestamp_ns => 1744208274700408014, :state => "waiting for gvl", :"thread id" => "10030 (29328)", :"thread name" => "/__w/dd-trace-rb/dd-trace-rb/spec/spec_helper.rb:276"} to include {end_timestamp_ns: (be between 1744208274700440756 and 1744208274700556813 (inclusive))}
       Diff:
       @@ -1,3 +1,5 @@
       -:end_timestamp_ns => (be between 1744208274700440756 and 1744208274700556813 (inclusive)),
       +:"thread id" => "10030 (29328)",
       +:"thread name" => "/__w/dd-trace-rb/dd-trace-rb/spec/spec_helper.rb:276",
       +:end_timestamp_ns => 1744208274700408014,
        :state => "waiting for gvl",
     # ./spec/datadog/profiling/collectors/thread_context_spec.rb:1205:in 'block (6 levels) in <top (required)>'
     # ./spec/spec_helper.rb:261:in 'block (2 levels) in <top (required)>'
     # ./spec/spec_helper.rb:146:in 'block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/webmock-3.24.0/lib/webmock/rspec.rb:39:in 'block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/rspec-wait-0.0.10/lib/rspec/wait.rb:47:in 'block (2 levels) in <top (required)>'

Finished in 20.91 seconds (files took 1.01 seconds to load)
764 examples, 1 failure, 19 pending

Failed examples:

rspec ./spec/datadog/profiling/collectors/thread_context_spec.rb:1196 # Datadog::Profiling::Collectors::ThreadContext#sample timeline support when timeline is enabled when thread starts Waiting for GVL records a second sample to represent the time spent Waiting for GVL

Randomized with seed 49054
```

We previously tried to solve this problem in #4495 but the previous
attempt was clearly not entirely successful.

The core of the issue is that internally the profiler uses a monotonic
clock when measuring wall-time, and then needs to turn it into system time.

Since there's no way (that I know of) to get both a monotonic and a
system clock timestamp at once, our conversion code gets both in
sequence and then uses that for the conversions.

Prior to this PR, our tests used
`Datadog::Core::Utils::Time.as_utc_epoch_ns(Time.now)` and then tried to
compare it with what the profiler saw.

This was the source of the flakiness: sometimes the conversion was
slightly off from the measurements we got when reading the system clock
directly.

For instance, here's an example of a situation where this can happen:
1. Test code reads system_clock = 100
2. Profiler code reads monotonic_clock = 2000
3. Profiler code wants to convert to system clock, and needs to
   initialize the conversion code.
  3.1 Profiler reads system_clock = 105
  3.2 Profiler reads monotonic_clock = 2010
  3.3 Profiler delta to epoch ns is 105-2010 => −1905
4. Profiler converts monotonic_clock = 2000 to system using delta:
  2000-1905 => 95
5. Test code expected 95 > 100 => failure

I suspect a variant of this is what happened in the test above.

To solve this flakiness once and for all, the tests were changed to
**use the same clock source** as the profiler, thus guaranteeing that if the tests
get a timestamp (call it t1), and then call the profiler and it gets a timestamp
(call it t2), that t2 >= t1 (and vice-versa).

**Motivation:**

Get profiler to zero flaky tests!

**Additional Notes:**

This isn't the only way to solve this issue... We could for instance
inject the timestamps from the tests to inside the profiler, which
would kind-of be the reverse of this solution: get the profiler to
use the Ruby-level test timestamps.

But, that would need a lot more plumbing and changing of production code
which doesn't seem like the right trade-off for this problem.

Finally we could always move the profiler to use system clock directly,
but that's a whole can of worms too, and arguably it wouldn't be any
better than it is now.

**How to test the change?**

The tests should be green and remain green and not be flaky!