Update `profiler_http_transport` benchmark to generate a new profile on every report
Reusing an `encoded_profile` object will no longer be possible with
libdatadog 17+.

This does change what the benchmark was covering, so comparing with
prior results is not particularly relevant.

I did consider removing this benchmark (we haven't used it for a
long time for anything in particular), but on the other hand these
microbenchmarks that exercise libdatadog have been useful sometimes so
why not.