[NO-TICKET] Minor: Make sure `datadog_ruby_common.h/.c` stay in sync
**What does this PR do?**

As per the comment we have on `datadog_ruby_common.h/.c`, to avoid
having too much of a weird setup in our build, we have a copy
of these files for profiling and libdatadog_api.

BUT rather than just having a comment saying "please remember to
keep these up-to-date", let's add a test to make sure that happens.

**Motivation:**

As I was reviewing a PR and was about to add a comment saying
"please keep these files in sync" I decided I would write some
code instead of just leaving a comment.

**Additional Notes:**

Suggestions are welcome on how to better solve this without
having a too-weird `extconf.rb` setup.

**How to test the change?**

I've tried changing the files on only one of the folders and
validated the test correctly failed.