[PROF-11680] Add support for profiling 3.5.0-preview1
**What does this PR do?**

This PR adds support for profiling Ruby 3.5.0-preview1. The only
work needed was importing the latest headers in
https://github.com/DataDog/datadog-ruby_core_source/pull/15
so this PR adds the new 3.4.1 release of the headers as the minimum
version allowed.

**Motivation:**

By setting the minimum version of `datadog-ruby_core_source` to the
latest release, we make sure bundler picks it up, and doesn't try
to use an incorrect release with Ruby 3.5.

**Additional Notes:**

N/A

**How to test the change?**

We don't yet test with 3.5 on CI; I tested locally that the profiler
test suite is working fine with 3.4 and 3.5.