Fix crashtracking broken due to too strict timeout
It took us a while to understand why crashtracking was not working
correctly with modern versions of libdatadog.

It turns out the problem was that we were setting a too low timeout
for Ruby crash reports.

I suspect this issue has been hidden for a while because before
libdatadog 17, crashtracking was almost always giving up on stack
symbolication ( https://github.com/DataDog/libdatadog/pull/991 ),
which was a much faster path, and thus we didn't notice
that the timeout required adjusting until libdatadog 17 came
around and fixed symbolization (making it slower).

I first tried with the libdatadog default of 5 seconds and it
still wasn't enough on my machine. Then I bumped it slightly
higher and that seems to be enough.

I'm surprised symbolication is this expensive...