Add extra sanity check to investigate flaky spec
This test was flaky in CI in the test-asan run:

https://github.com/DataDog/dd-trace-rb/actions/runs/14973937437/job/42061305617?pr=4638
https://github.com/DataDog/dd-trace-rb/actions/runs/14697248627/job/41240596719?pr=4588

```
  1) Datadog::Profiling::StackRecorder#serialize heap samples and sizes
     when enabled #recorder_after_gc_step when
     heap_clean_after_gc_enabled is true clears young dead objects with
     age 1 and 2, but not older objects

     Failure/Error: expect(@object_ids.map { |it| is_object_recorded?(it) }).to eq [false, false, false, false]

       expected: [false, false, false, false]
            got: [true, false, false, false]
```

This result is extremely weird/suspect, because it looks like the
Ruby GC has cleared 3 out of 4 of the objects that were supposed to
be cleared.

I'm not sure why this is happening (and is it related to the ASAN
build?) so I'm adding an extra test to check the assumption that
all objects have been correctly cleared.