[NO-TICKET] Fix profiling scheduler reporting corner case during shutdown
**What does this PR do?**

This PR fixes a weird corner case in the profiling scheduler class that
happened when the scheduler was being shutdown / stopped.

Specifically, if reporting a given profile to the backend took longer
than `Exporter::PROFILE_DURATION_THRESHOLD_SECONDS`, then the
exporter would keep on returning `exporter.can_flush? == true`.

This caused the scheduler to get "stuck" in a loop trying to report
profiles again and again, until one of the reports was fast enough
so that `exporter.can_flush? == false`.

In practice, this wouldn't actually happen because during shutdown,
the `stop` operation gets called with `force_stop = true`
(in workers/polling.rb) and thus the scheduler would be terminated
"the hard way" after `DEFAULT_SHUTDOWN_TIMEOUT`, which is set to
1 second.

**Motivation:**

Even though this wouldn't happen in practice, it was a footgun locked
and loaded.

This footgun actually hit me while I was doing some unrelated
experiments and raised `DEFAULT_SHUTDOWN_TIMEOUT` to test something out.

So I decided to disarm the footgun and add test coverage for it ;)

**Additional Notes:**

N/A

**How to test the change?**

I've added test coverage for the change. If you want to test it
manually, try setting `Exporter::PROFILE_DURATION_THRESHOLD_SECONDS`
to a really low value, such as 0.01 and observe what happens on
app shutdown.