[NO-TICKET] Update crashtracking specs for next libdatadog release
**What does this PR do?**

This PR prepares crashtracking spec changes needed to support two
recent changes to libdatadog:

1. https://github.com/DataDog/libdatadog/pull/1037
2. https://github.com/DataDog/libdatadog/pull/1076

With this PR I'm able to get 100% green run for profiling and
crashtracking with libdatadog master.

**Motivation:**

Get us in good shape for the next libdatadog release.

**Additional Notes:**

I'm opening this PR as a draft as we're still discussing one of
the behavior changes.

Once that gets decided, I'll either remove just the TODO note or
I'll undo that part of the change.

*How to test the change?**

If CI is green, we're good! This change is actually backwards-compatible
so it's still compatible with the libdatadog v18.1.0 release that
we're using right now for Ruby.