[NO-TICKET] Review and clean up profiler logging
**What does this PR do?**

This PR does a pass on the profiler's logging and:
* Lowers the severity of a few messages from error to warn
* Reports a few more messages via telemetry
* Changes a few of the telemetry reports to also include exception
  information

**Motivation:**

Recently I was investigating an issue in our reliability environment
and I missed some of the logging I wanted, because it wasn't being
reported using telemetry. So I decided to do that cleanup.

Regarding the error to warn changes, since the profiler failing is not
in any way expected to impact the application, it seems to me that it's
clearer that such failures are warnings ("hey -- something went wrong
and you may not get any profiles").

**Additional Notes:**

N/A

**How to test the change?**

I've updated our unit specs where needed, so we should be good to go.