Do not use native filename when top of stack is Ruby VM frame
Using the ruby_native_filename at the top of the stack has a weird
effect on the "top methods" table because e.g. we don't have classnames
for methods.

This is especially visible in the allocations profile, e.g. what a
surprise, you're telling me "libruby.so:new" is the top method always?

Until we have a better way of dealing with that, we don't do this
replacement for the top frame.