Tweak setup for thread with deep stack
The target depth was not actually what we were getting for
`thread_with_very_deep_stack_and_native_frames` we have 2 extra frames
per call.

To fix this, just look at the actual stack depth rather than trying to
count.

This needed to be changed as under Ruby 2.5 we were hitting a stack
overflow due to the very deep stacks we were getting in the
benchmark.