Introduce `thread_context_collector_prepare_sample_inside_signal_handler` + `prepare_sample_thread`
These APIs are the key pieces to enabling sampling from the signal
handler: they allow the actual sampling of the stack to be done
into the sampling buffer, but the buffer is not flushed until the next
regular sample.

Note that
`thread_context_collector_prepare_sample_inside_signal_handler` still
needs to be wired into the `CpuAndWallTimeWorker` as right now it only
gets triggered in the tests.