Don't run code provenance specs on JRuby
Technically they could be made to run, but no use adding extra
complexity to support JRuby when the whole profiler does not support
JRuby.

Only specs where there's 0 cost in running in JRuby are worth it (and in
that case we're basically getting a second opinion from JRuby for "free"
on whatever is pure-Ruby).