[PROF-12098] Profiler: Sample from inside signal handler on modern Rubies
**What does this PR do?**

In https://github.com/DataDog/dd-trace-rb/pull/4785 we added a setting
to enable sampling from inside the signal handler, but made it off by
default.

The plan was to wait for validation in our test apps before flipping
the default. The test apps have now been running for a while with this
feature without issue, so we're ready to enable it.

This PR changes the setting to be on by default on Ruby 3.2.5+/3.3.4+.

**Motivation:**

The Ruby versions listed include https://github.com/ruby/ruby/pull/11036
(as well as other prior fixes) and thus it's safe to use this feature,
which improves accuracy of profiling data.

See original PR for more details.

**Additional Notes:**

Out of curiosity, our test apps use Ruby 3.2.5 (GitLab) and 3.3.4
(Rails) so they are exactly at the right versions so they'll start
getting the feature on by default.

As mentioned in the comment attached to the setting, we could
probably enable this feature on more Ruby versions (all of the 3.x
series, potentially) but with more than half of all customer
accounts using these modern Rubies, I'm not sure it's worth the
extra validation work at this point.

**How to test the change?**

The change includes test coverage. Furthermore, when sampling from
the signal handler is enabled, the `signal_handler_prepared_sample`
counter added in https://github.com/DataDog/dd-trace-rb/pull/4785
will be non-zero, so this can be used as an extra validation that
the feature is enabled for a given profile.