[NO-TICKET] Profiling: Add potential fix for Ruby VM bug
**What does this PR do?**

This PR tries to import the fix added in
https://github.com/ruby/ruby/pull/13643 / https://bugs.ruby-lang.org/issues/21441
by applying it on the "reader" side: e.g. assuming we may see
`ec != NULL` but `vm_stack == NULL` or `vm_stack_size == 0` during
sampling.

**Motivation:**

I am not sure if this could happen in practice for the profiler, but
given the extra check is cheap and means that our version of the fix
applies to all Rubies without any need for the upstream fix, this
seems worth to have.

**Additional Notes:**

N/A

**How to test the change?**

Unfortunately I don't have a lot of ideas for how to test this :/

(Other than spending some amount of time to create a throwaway modified
Ruby that would trigger profiling on the exact situation and
see if breaks or not... Given that's quite an investment and the
change seems quite simple, I'm not sure it's worth the time.)