[PROF-11475] Profiling: Improve and tweak crash workaround
**What does this PR do?**

In https://github.com/DataDog/dd-trace-rb/pull/4464 we added a few
extra checks to `calc_pos` as an attempt of fixing an issue, but at
the time we weren't sure of the exact root cause of the issue.

Recently we got a second crash report from an old library version
without #4464 and with this extra info I was finally able to
understand and reproduce the issue.

Thus I can now confirm the issue is with `n` going out of bounds
and thus our attempted fix was correct which is probably we got
exactly only 2 reports of this crash.

**Motivation:**

Update our documentation of this issue and remove the
unneeded `RB_TYPE_P(iseq, T_IMEMO)` check.

**Additional Notes:**

N/A

**How to test the change?**

See
<https://docs.google.com/document/d/1I0edOVmLRDCdjdgKBy_dZTJHwa04781xvc1XGq3BOyQ/>
(Datadog-only link, sorry!) for my notes on reproducing the issue.