[NO-TICKET] Tag profiles with sequence number
**What does this PR do?**

This PR adds a new tag to reported profiles: `profile_seq`.
This tag is provides a sequence number for reported profiles for a
given runtime-id.

**Motivation:**

This tag has already been adopted by the Go and .NET profilers, and is
useful for use-cases such as:

* Focus on, or exclude application boot time (which usually has quite
  different profiles)
* Clearly identify when profiles are missing (because e.g. they
  failed to be reported and there's gaps in the sequence)

**Additional Notes:**

This feature is quite straightforward, the only weird part is that
we carry state across profiler reconfigurations (e.g. state is
a singleton), and we also need to reset it after the app forks,
because runtime-id changes.

**How to test the change?**

This change includes test coverage.