[PROF-12141] Upgrade to libdatadog 19.1
**What does this PR do?**

This PR bumps the libdatadog version to 19.1 --
https://github.com/DataDog/libdatadog/pull/1146 .

**Motivation:**

Make sure we're up-to-date with the latest libdatadog goodies.
In particular, this version brings faster crashtracking reporting
(which is nice because otherwise it could time out and not report
data, or only report partial data).

**Additional Notes:**

N/A

**How to test the change?**

We have test coverage for features that depend on libdatadog, so
green CI is good!