[PROF-12143] Report GC tuning environment variables with profiles
**What does this PR do?**

This PR tweaks the information that gets attached to every profile
to include any GC tuning settings that were present.

This information can be seen in the "System Info" section of the
single profile view.

**Motivation:**

As more GC tuning settings get added to Ruby, it's useful for
performance reasons to know if a given app is operating with the
defaults or something has been tweaked.

See <https://mailchi.mp/railsspeed/stop-tuning-rubys-gc> for an
interesting discussion of this.

We want to make it easier to spot these GC tuning settings
(maybe even automatically in the future?).

**Additional Notes:**

N/A

**How to test the change?**

I've added test coverage, and plan to attach a screenshot showing
this information for a test app.