[NO-TICKET] Retire `error_tracking_api` benchmark
**What does this PR do?**

This PR retires (removes) the `error_tracking_api` benchmark until
we can have more time to fix it.

**Motivation:**

This benchmark was causing some flakiness and in
https://github.com/DataDog/dd-trace-rb/pull/4815 as we were
tweaking the code we uncovered that actually the benchmark wasn't very
sensitive to any changes in error tracking since the `sleep`s we
were using were too large for the thing being measured
(see
https://github.com/DataDog/dd-trace-rb/pull/4815#discussion_r2216009394).

I discussed with @dubloom via chat, and he won't be able to resolve
this bigger issue of the benchmark -- which is still blocking CI, so
we agreed to remove until it can be later restored and fixed.

Thus, this PR replaces https://github.com/DataDog/dd-trace-rb/pull/4815 .

**Additional Notes:**

N/A

**How to test the change?**

Validate that CI is green!