Reverse stack sampling direction to improve stack cache reuse hits
For a while we've had a very simple caching mechanism where if we
detect that a frame being written into the `stack_buffer` is the
same as before, we avoid redoing some of the same work.

But because the `rb_profile_frames` function we started from samples
the stack starting from the leaf, if between two samples the stack
depth changed (e.g. we've entered or left a function), then the caching
would no longer work, because e.g. the frame that was previously
sampled at position 1 is now seen at position 0, and so on for the
rest of the stack.

To better take advantage of caching, let's iterate the stack in
the other direction: starting from the root (oldest frame) towards
the leaf (current function). This means that most of the `stack_buffer`
remains stable when there's just changes going on in the leafs.

This matches what would happen in, e.g. a web request: the path
through the webserver and to get to the endpoint would remain
stable for most of the request, and can now be cached.

There is one non-trivial consequence for this: frame truncation
now shows the root of the thread and not the leaves. This
behavior is closer to what the python profiler is doing, and
also avoids mismatched stacks from doing the truncation in
the other direction.

(Note that right now, unlike Python, the truncation frame is
being placed at the root and not the leaf -- I'll fix that
next)