Fix off-by-one in stack depth and remove `stack_depth_for`
The `stack_depth_for` helper was giving off-by-one values, and we didn't
notice as the

```ruby
Thread.new(&DeepStackSimulator.new(target_depth: depth, ready_queue: ready_queue).method(first_method))
```

in the tests introduces a special kind of frame that gets skipped
when sampling the stack and thus made the values match for the test.

That is: `stack_depth_for` would give 5 for a depth of 6, but
we got "lucky" that one of those frames got skipped over and thus
we would get 5 frames out of sampling. But when we turn that
frame into a regular one (by not using `&...method(...)`) the
off-by-one would be uncovered.

Having said that, rather than trying to fix `stack_depth_for`, which
to be completely correct it should be a mini-`rb_profile_frames`
without recording the observed info, let's get rid of it.

This does mean that if the stack depth == maximum depth we
still truncate 1, but that seems fine since the depth can always
be configured up as needed to compensate, so I think the simplified
logic is worth it.