Tweak `profiling_sample_loop_v2.rb` benchmark to reduce complexity in test suite
We're only trying to validate the benchmark isn't broken due to API
changes so I think it's fine to run with a much smaller depth.

We saw this benchmark fail at least once in CI, presumably because
the `go_to_depth_and_run` took some time to setup (in my machine I
needed more than 1 second to run this benchmark in validation mode).