Hack: Make sure runtime-id is correct on forks and reconfiguration
The semantics for `shutdown!` on reconfiguration are a bit different
from what I expected so the code to drop the context wasn't quite
correct. In particular, when reconfiguration happens, the new component
gets started before the older one shuts down; this is fine generally
but since the context is a singleton it means my approach of dropping
on `shutdown!` was not correct as it was tearing down the context
after updating it.

As a simplification, let's never drop the context.

+ Also fix updating the runtime-id on forks; the process discovery
module actually is incorrect as it wasn't handling this, and so I
did a bit of a heavy-handed thingy to fix.