Improve reading output from forked process
A few small tweaks:

1. I noticed that without `$stderr.sync = true` we often missed the actual
stderr output, if any.

See [this link](https://stackoverflow.com/questions/3018595/how-do-i-redirect-stderr-and-stdout-to-file-for-a-ruby-script)
for a tiny discussion of this.

2. While reading the documentation for `TempFile` it also mentioned

> The underlying file is removed when the Tempfile object dies and is
> reclaimed by the garbage collector.

So I moved the close around "just in case" to make sure the file
never got removed too early.

3. I replaced the `||=` with a `=` when reading the files also
"just in case" so that we read the latest content of the file rather
than try to reuse content that might be slightly stale.